Remove wrong int64 cast on the disk resources.

+++ command/agent/fs_endpoint_test.go
@@ -449,7 +449,7 @@ func tempAllocDir(t *testing.T) *allocdir.AllocDir {
 		t.Fatalf("failed to chmod dir: %v", err)
 	}
 
-	return allocdir.NewAllocDir(dir, int64(structs.DefaultResources().DiskMB))
+	return allocdir.NewAllocDir(dir, structs.DefaultResources().DiskMB)
 }
 
 type nopWriteCloser struct {
